requirements.txt
pysqlite3-binary
python-dotenv
streamlit
sentence-transformers
unstructured
langchain
langchain-community
langchain-google-genai
langchain-experimental
langchainhub
faiss-cpu